+++ go.mod
@@ -59,6 +59,6 @@ require (
 	google.golang.org/protobuf v1.28.0 // indirect
 )
 
-replace github.com/vmware/go-vcloud-director/v2 => github.com/Didainius/go-vcloud-director/v2 v2.17.0-alpha.2.0.20220831192258-1eb0851f5eb3
+replace github.com/vmware/go-vcloud-director/v2 => github.com/Didainius/go-vcloud-director/v2 v2.17.0-alpha.2.0.20220923120331-3234e09d238c
 
 //replace github.com/vmware/go-vcloud-director/v2 => ../go-vcloud-director
